Move files into stats folder
