+++ src/optimizer/stats/stats.cpp
@@ -2,25 +2,22 @@
 //
 //                         Peloton
 //
-// tuple_sample.cpp
+// stats.cpp
 //
-// Identification: src/optimizer/tuple_sample.cpp
+// Identification: src/optimizer/stats.cpp
 //
 // Copyright (c) 2015-16, Carnegie Mellon University Database Group
 //
 //===----------------------------------------------------------------------===//
 
-#include "optimizer/tuple_sample.h"
+#include "optimizer/stats/stats.h"
 
 namespace peloton {
 namespace optimizer {
 
 //===--------------------------------------------------------------------===//
-// TupleSample
+// Stats
 //===--------------------------------------------------------------------===//
-TupleSample::TupleSample(std::vector<Column *> columns,
-                         storage::TileGroup *sampled_tuples)
-    : columns(columns), sampled_tuples(sampled_tuples) {}
 
 }  // namespace optimizer
 }  // namespace peloton
